[6/N] Extract and reuse utility functions in DockerBuiler
