+++ sematic/config/settings.py
@@ -1,5 +1,4 @@
 # Standard Library
-import distutils.util
 import enum
 import logging
 import os
@@ -306,23 +305,6 @@ class EnumDumper(yaml.Dumper):
 
     def ignore_aliases(self, data: Any) -> bool:
         return True
-
-
-def as_bool(value: Optional[Any]) -> bool:
-    """
-    Returns a boolean interpretation of the contents of the specified value.
-    """
-    if isinstance(value, bool):
-        return value
-
-    if value is None:
-        return False
-
-    str_value = str(value)
-    if len(str_value) == 0:
-        return False
-
-    return bool(distutils.util.strtobool(str_value))
 
 
 def _load_settings(file_path: str) -> Settings:
