+++ sematic/resolvers/tests/test_worker.py
@@ -35,11 +35,8 @@ from sematic.function import func
 from sematic.future_context import PrivateContext, SematicContext
 from sematic.resolvers.cloud_resolver import CloudResolver
 from sematic.resolvers.worker import _emulate_interpreter, main, wrap_main_with_logging
-from sematic.tests.fixtures import (  # noqa: F401
-    environment_variables,
-    test_storage,
-    valid_client_version,
-)
+from sematic.tests.fixtures import test_storage, valid_client_version  # noqa: F401
+from sematic.utils.env import environment_variables
 from sematic.utils.stdout import redirect_to_file
